# src/main/resources/openapi/short-url-api.yaml
openapi: 3.1.0

info:
  title: Short URL Service API
  version: 1.0.0
  description: >
    Public and administrative APIs for creating, resolving and managing short URLs.
servers:
  - url: http://localhost:8080
tags:
  - name: Short URL
    description: Public URL operations
  - name: Admin URL
    description: Administrative URL management
paths:
  /{code}:
    get:
      tags:
        - Short URL
      summary: Redirect to original URL
      description: Redirects the client to the original URL.
      operationId: redirect
      parameters:
        - name: code
          in: path
          required: true
          description: Short URL code
          schema:
            type: string
      responses:
        "302":
          description: Redirect to original URL
          headers:
            Location:
              description: Destination URL
              schema:
                type: string
                format: uri
        "404":
          description: Short URL not found

  /api/v1/urls:
    post:
      tags:
        - Admin URL
      summary: Create Short URL
      operationId: createShortUrl
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShortUrlRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortUrlResponse'

  /api/v1/urls/{code}:
    get:
      tags:
        - Admin URL
      summary: Get URL Details
      operationId: getShortUrl
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/Code'
      responses:
        "200":
          description: URL found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortUrlResponse'
        "404":
          description: Not Found

    put:
      tags:
        - Admin URL
      summary: Update Short URL
      operationId: updateShortUrl
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/Code'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShortUrlRequest'
      responses:
        "200":
          description: Updated

    delete:
      tags:
        - Admin URL
      summary: Delete Short URL
      operationId: deleteShortUrl
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/Code'
      responses:
        "204":
          description: Deleted

  /api/v1/urls/{code}/stats:
    get:
      tags:
        - Admin URL
      summary: URL Analytics
      operationId: getStatistics
      security:
        - bearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/Code'
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'

components:
  parameters:
    Code:
      name: code
      in: path
      required: true
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    CreateShortUrlRequest:
      type: object
      required:
        - longUrl
      properties:
        longUrl:
          type: string
          format: uri
          example: https://openai.com
        customAlias:
          type: string
          example: openai

    UpdateShortUrlRequest:
      type: object
      properties:
        longUrl:
          type: string
          format: uri
        enabled:
          type: boolean

    ShortUrlResponse:
      type: object
      properties:
        code:
          type: string
          example: Ab3XkL
        shortUrl:
          type: string
          example: http://localhost:8080/Ab3XkL
        longUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        clickCount:
          type: integer
        enabled:
          type: boolean

    StatisticsResponse:
      type: object
      properties:
        code:
          type: string
        totalClicks:
          type: integer
        uniqueVisitors:
          type: integer
        lastAccessed:
          type: string
          format: date-time
        topCountries:
          type: array
          items:
            type: string